Make Redux store persistent, fix async thunk
Some settings like position and nickname should persist between app
sessions – that is, if you close the app and open it again, you should
see the same area of the map and be able to comment with the same
nickname as before.

There was no real issue with the actual thunk code, the problem was
rather the way I called it and the fact that only serializable objects
can be stored in a Redux store – no instances of the Dumpster class, for
example. This will have consequences for later stages of development.