Fix bug in dumpster DAO
Store and dumpster types were returned as undefined values when a
dumpster was posted or updated. This was fixed by using the values _sent
in_ with the request. Hooray.

However, this does not fix everything. The rating that is returned, is
the rating you'd receive if none existed. A simple 2.5 points. This
works for the POST request, since no rating has been given yet, but
absolutely not for edited dumpsters.

Conclusion: Consider _either_ adding more work to the backend (fetching
rating when a dumpster is updated) _or_ adding even more work by having
the frontend request the updated dumpster after a successful PUT
request. Pick your poison.