Clean up dumpster route
Removed stuff that didn't match the current API specification or just
was too much of a mess to keep. It'd have to be changed anyway.
Also made sure that the overly simple routes like /categories actually
send errors to the next function. At the moment they'd potentially just
crash, because Express does not automatically send stuff to next() when
async functions fail (double check this, it'd be nice if it did).