Make endpoint indicate which revision is active
Added a simple isActive boolean.
I'm starting to think that Jon's solution might actually be better
here... Having a boolean that indicates that a revision is the active
one would make this and *many other things* much simpler. However, I
think there would be some (unforeseen) cost to doing so.