Add input validation to dumpster editor
I chose to go for Formik. Might want to consider switching to Hook Form
for greater speed. In any case, the biggest pain for us developers is
the fact that both libraries seem to exclusively deal with strings. It's
rather cumbersome, as you may see from some of this (messy) code.