Add intro text and language picker on intro screen
I also did something that fixed the bug where a reload triggered by Expo
would cause the language to reset to English. A simple if (language)
saved the day.

However, there are some wonky problems with the intro page styling.
If more languages are added, the list will overflow. So sad.